+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add lock icon
 - Add `forceHeader` prop to `Input` component
+- Added `confirmButtonIsLoading` prop on `<Modal />`
+- Added a `<SelectInput />` component
 
 ### Fixed
 - Switcher: better style for disabled state